Change double to float.

+++ Runtime/animation/animation.cs
@@ -59,7 +59,7 @@ namespace Unity.UIWidgets.animation {
 
         public Animation<U> drive<U>(Animatable<U> child) {
             D.assert(this is Animation<double>);
-            return child.animate(this as Animation<double>);
+            return child.animate(this as Animation<float>);
         }
     }
 }
