+++ Runtime/gestures/converter.cs
@@ -62,7 +62,7 @@ namespace Unity.UIWidgets.gestures {
                 () => new _PointerState(position));
         }
 
-        public static IEnumerable<PointerEvent> expand(IEnumerable<PointerData> data, double devicePixelRatio) {
+        public static IEnumerable<PointerEvent> expand(IEnumerable<PointerData> data, float devicePixelRatio) {
             foreach (PointerData datum in data) {
                 var position = new Offset(datum.physicalX, datum.physicalY) / devicePixelRatio;
                 var timeStamp = datum.timeStamp;
